feature: Post service Test code refactor: Category Id type inconsistency problem

+++ src/main/java/com/junhee/spring_board_api/domain/post/dto/PostUpdateRequest.java
@@ -7,7 +7,7 @@ import lombok.Setter;
 @Getter
 @Setter
 public class PostUpdateRequest {
-    private Long categoryId;
+    private int categoryId;
     @NotBlank(message = "제목은 필수입니다.")
     private String title;
     @NotBlank(message = "내용은 필수입니다.")
